Tighten temporal warning feedback handling
